Make keywords and symbols first-class token types.
It's hard to check for a specific keyword or symbol without reading
its string from the file again, which makes parsing slightly more
complicated.

While it's possible to just include the string in the token structure
itself, it means that we have to dynamically allocate and free lots
of (potentially duplicate) strings in the token structure, which is
slower and more error-prone.

This solution is more performant, saves memory and gets rid of two
global variables, which are all great benefits.